Fix bad int -> double conversion on division

+++ Project/src/ColorSimilarityMeasures.cpp
@@ -7,10 +7,14 @@ std::pair<double, double> **ColorSimilarityMeasures(Color **img1, Color **img2, 
     std::pair<double, double> **retVal = new std::pair<double, double> *[width];
     for (int i = 0; i < width; i++)
         retVal[i] = new std::pair<double, double>[height];
+
     for (int x = 0; x < width; x++) {
         for (int y = 0; y < height; y++) {
-            double r = std::min(img1[x][y].r, img2[x][y].r) / std::max(img1[x][y].r, img2[x][y].r);
-            double g = std::min(img1[x][y].g, img2[x][y].g) / std::max(img1[x][y].g, img2[x][y].g);
+            double r = (double) std::min(img1[x][y].r, img2[x][y].r) / (double) std::max(img1[x][y].r, img2[x][y].r);
+
+			//std::cout << "r1:" << (int) img1[x][y].r << " r2:" << (int) img2[x][y].r << " frac_r: " << r << std::endl;
+
+			double g = (double) std::min(img1[x][y].g, img2[x][y].g) / (double) std::max(img1[x][y].g, img2[x][y].g);
             retVal[x][y] = std::make_pair(r, g);
         }
     }
